move `MessageState` type to package
