+++ packages/types/src/definitions/runtime/MessageControl.type.ts
@@ -7,3 +7,10 @@ export type TimerMessage = Message & {
   blink: boolean;
   blackout: boolean;
 };
+
+export type MessageState = {
+  timer: TimerMessage;
+  public: Message;
+  lower: Message;
+  external: Message;
+};
