+++ packages/types/src/definitions/runtime/RuntimeStore.type.ts
@@ -1,5 +1,5 @@
 import { Playback } from './Playback.type.js';
-import { Message, TimerMessage } from './MessageControl.type.js';
+import { MessageState } from './MessageControl.type.js';
 import { TimerState } from './TimerState.type.js';
 import { Runtime } from './Runtime.type.js';
 import { OntimeEvent } from '../core/OntimeEvent.type.js';
@@ -10,12 +10,7 @@ export type RuntimeStore = {
   playback: Playback;
 
   // messages service
-  message: {
-    timer: TimerMessage;
-    public: Message;
-    lower: Message;
-    external: Message;
-  };
+  message: MessageState;
   onAir: boolean;
 
   // event loader
